:sparkles: no more profile controller
