--- src/services/profile-thunks.js
@@ -1,16 +1,0 @@
-import { createAsyncThunk } from '@reduxjs/toolkit';
-import * as profileService from "./profile-service";
-
-export const fetchProfile = createAsyncThunk(
-  "profile/fetchProfile", async (profileId) => {
-    const profile = await profileService.getProfile(profileId);
-    return profile;
-  }
-);
-
-export const updateProfile = createAsyncThunk(
-  "profile/updateProfile", async ({id, profileUpdate}) => {
-    const updatedProfile = await profileService.updateProfile(id, profileUpdate);
-    return updatedProfile;
-  }
-);
